Set the minimum Python3 to 3.3
This was not really set. The setup supported 3.2, the default was saying 3.4,
and I have set it now to 3.3.

Reason:
Besides a sensible useful minimum version, Python 3.3.3 is the only currently
working version, which allows for a debug python build on macOS. I need this
for my debugging work, admittedly.
So this change is really made to support PYSIDE-79.

Victor Stinner has supplied a patch three years ago, see

https://mail.python.org/pipermail/python-bugs-list/2013-July/209044.html

Since then, the patch was not maintained, and the problem has crept into python, again.
If I find some time for it, then I will fix it.

Task-number: PYSIDE-79
Change-Id: I1b7d4ff1613d31567095636edc61f4222fad08dd
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>