Fix Segfault when using connect (w/o instrumentation)
The multiple problems were caused by wrong or missing refcounts.
They led probably to the assumption that we need the garbage collector.

There is no garbage collector needed, since there exist no cyclic references.
Some increfs and decrefs were not correct, and a few special rules concerning
reference counting were ignored.

This is a final solution:

The case of a function call that returns a signal is special.
In this case, we keep track of the extra reference in a structure.
In all other cases there is a pre-existing object.

This patch is not creating any other errors.
It is no longer a hack but a true solution.

Added explicit path setting in order to circumvent the spurious python 2.7
error in disconnect_test.py .

Task-number: PYSIDE-79
Change-Id: I2579adf8fc671a2c2b665cfddaa26ecf96300a99
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>