Sanity check dynamically added signal and slots
Whenever a connection is made to a valid method that was not yet
defined as a slot, the method will be marked as a slot before the
connection proceeds, thus the connection will succeed.

However when connecting using old style SLOT("method()") syntax,
it is possible to provide an invalid signature by specifying the
method name, but forgetting the parentheses.
In such a case, instead of failing and notifying the user, the
connection erroneously succeeded, and on signal emission would not
behave correctly.

This change makes sure to sanity check that dynamically added
signals and slots contain parentheses, thus not causing incorrect
connection creation.

Change-Id: Iaa1a85a2ce698ca31a99100e00b683cfa0cb33cc
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>