Correctly print error message in case of infinite recursion
When a method is invoked recursively multiple times, it can exhaust
Python's stack depth limit, in which case an exception regarding that
limit is printed.

However the error message is not printed correctly, because
PyErr_Print itself tries to call Python code to normalize exceptions
if needed, which hits the stack depth limit issue again, and tries to
set the exception again, and fails.

This change makes sure to temporarily increase the stack limit, to allow
printing the exception message which says that the stack limit has been
reached.

Change-Id: I8eb086f9ad5b8712d348b6e894c2ef308f034528
Reviewed-by: Christian Tismer <tismer@stackless.com>