setup.py: Clean up version handling and options
OPTION_VERSION and OPTION_LISTVERSIONS are deprecated and not used
anymore. Because the shiboken and pyside2 submodules were merged into
the pyside-setup supermodule, it is not correct to advertise available
versions based on remote branches.

If a user provides --version 5.9 to a 5.6 based setup.py, the build
would fail because setup.py wouldn't handle clang specific options.

And if a user provides a --version 5.6.5 to a 5.6 based setup.py,
we can't re-checkout the current active repo to re-run an updated
setup.py.

Thus the version selection mechanism is no longer relevant, and it will
be strictly tied to branch / tag.

In this respect, some of the version handling code is removed (because
it wouldn't work) and cleaned up.

Task-number: PYSIDE-431
Change-Id: I8357d743b5e3a638cea583a763bffc7766688030
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>