Support multiple builds in testrunner.py
The testrunner module allowed only the last build to be tested.
This patch allows to specify an arbitrary test index.

The default is -1, which gives the normal behavior.
The argument ‘—buildno’ specifies which build is to be selected.

Note that multiple builds with the same directory are not valid, because the
latest entry always overwrites the older ones. But for different configurations
this makes sense.

The default runtime was set to 600 s.

Change-Id: Ib5b33a09dc0af7e36230ca4eba3c21c8970d8212
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>