Fix shiboken crash when generating template argument conversion code
Shiboken tried to generate conversion code for every argument in
a C++ method, and in case of a template method, it would crash
with an error because it couldn't resolve the type of the templated
argument when trying look it up in the type database. e.g.
template <class T> void foo(QList<T *>); would crash because T is not
a valid type.

Fix makes sure to skip and warn about any container type
instantiations which have unresolved types as part of their signature.

Change-Id: I34ac13dd83572ab9a7185ae7991276396e21b86e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>