Fix C++ parser to identify function pointers in parameter clauses
Previously function pointers inside function parameter clauses
were not recognized by the parser, and led to generation of incorrect
type information for the parameters.

Fix checks if a parameter declarator contains a top level
additional parameter clause, which implies that the declarator is
a function pointer. This will not work for nested function pointer
declarators though.

This does not fix generation of bindings regarding function pointers,
but is just a step forward to provide proper type information at the
parser level.

Change-Id: I6930414c06ef8e4cddbe5d67623b754f29f17025
Task-number: PYSIDE-455
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>