Fix crash when creating signatures of function pointers in a parameter
Previously any presence of a function pointer parameter (which had a
void return type) in a function declaration would cause a crash,
because the type of the whole function pointer was misinterpreted
as just "void", and when type() was called on such a parameter a null
pointer was returned instead of an AbstractMetaType object, which when
accessed caused a crash.

Fix consists in checking that the parameters of a function do not have
a void type (unless there's only one parameter).
Some additionals checks are added when trying to access the signature
of a function.

Change-Id: I4baa477732ce1e7c8a494137f37f58bc0df38a07
Task-number: PYSIDE-455
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>