Add additional numeric overflow checks
This patch addresses the missing overflow warnings for each
assertRaises found in overflow_test.py and in
implicitconv_numerical_test.py, specifically for the cases of big
numbers that can't be represented in long long's.

The patch doesn't cover all the possible cases, but the overflow checks
should be more robust now, especially in cases when the overflows
happened silently and returned -1 without showing any warnings.

Change-Id: Ifded579f5c11d4ae78d91f63374dd62c8cbf953f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>