Refactor TemplateInstance::expandCode()
Ensure a newline is added after the opening comment, right-trim
the code and add a newline before the closing comment.
This prevents code from being lost when writing a template in a single line:
<template name="name">code</template name>.
Adapt the test to cover both cases.

Use hash iterators instead of iterating over QHash::keys(),
avoiding the construction of a temporary container.

Task-number: PYSIDE-323
Change-Id: I7116f4284b16947242357bf6aee963a29f1fb16d
Reviewed-by: Christian Tismer <tismer@stackless.com>