Remove CodeModelPointer
Replace the QAtomicPointer-based CodeModelPointer by a QSharedPointer and
remove inheritance of the _CodeModelItem whose hierarchy they manage of
QSharedData. Remove all the _cast functions and use the functionality
provided by QSharedPointer instead.

It is not clear what the intention of the original design was as there
are no threads involved in parsing; but it was massively leaking memory,
which QSharedPointer automatically fixes.

Task-number: PYSIDE-323
Change-Id: Icb1f265a14153697b8f9681dd0a1ab2e3af06176
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>