Fix “Fix Segfault when using connect”
There has been an over-done correction in this issue.
The issue was solved, but I also “corrected” the function “getFromType”:

- I added a check for a valid object. This was wrong, it is allowed to return
  a null object.

- I adjusted the refcount, because PyDict_GetItem returns a borrowed ref.
  But this was correct, since the caller added a reference.

I have run all tests. The result is identical before and after my change.

It was necessary to treat this case before the end of PYSIDE-315, because
the new test code from Mingxiang of today hits exactly this bug.

Task-number: PYSIDE-79
Task-number: PYSIDE-315
Change-Id: Id489e106a5ee7410ed32b5c4bd21d4b98e4e97a7
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>