Warn if signals and slots are out of order
The last improvement to signals and slots registration has removed
the problem that mixin classes were not handled the same as other classes.
The key idea was to use the MRO to inspect all involved classes
at type parsing time.

The signals and slots were then sorted in ‘updateMetaObject’.

The current patch enhances this solution in the following way:

In ‘parsePythonType’ we re-ordered the introspection loop in a way
that now all signals are extracted first. This way, they are ordered before
slots automatically, and a later sorting is not necessary. Furthermore,
instead of using qStableSort in ‘updateMetaObject’, we now check the
sorting only with ‘is_sorted’. If the sort order becomes invalid, it will issue a warning.

The latest change removes quadratic time behavior from the warning output.

Important notes and implications:
It is no longer relevant if slots are decorated with @QtCore.Slot().
Signals will work in normally created classes of all shapes.
If classes are modified later, they may grow an incompatibility
when signals appear after slots. Then you get a warning.

Adendum: It makes sense to use slots whenever possible. This way,
constructions as in static_metaobject_test.py work.

Now handling the warning correctly when it is turned into an error.
Correction: We leave the warning as-is and ignore the error!

Question: The static_metaobject_test.py passes its tests suddenly, when you use @Slot() !
Why? Should we open an extra issue for that?

Task-number: PYSIDE-315
Change-Id: I75c9c88787cd93251b1439db0088cd66fc0c3c97
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>