AbstractMetaBuilder: Fix type check
AbstractMetaBuilderPrivate::fixReturnTypeOfConversionOperator()
is supposed to apply some changes to a conversion operator
like

class A {
   operator int() const;
};

The old C++ parser has a bug in that it creates this as a function
of return type "A" instead of "int" and this is also checked at the top
of the function.

Remove the check since that causes the function to not do anything
if the correct type "int" is passed.

Task-number: PYSIDE-323
Change-Id: I19e5b83599dccd5737209dd2e92064e80e8a506d
Reviewed-by: Christian Tismer <tismer@stackless.com>