Fix emission of signals with default arguments
Previously when a signal was emitted by omitting default arguments,
PySide would complain that not enough arguments were given, because
only one of the signal overloads was considered for emission.

Fix consists in trying to find a signal overload which has the same
number of arguments as given to the emit statement, as well as being
marked as a cloned meta method (which for moc means a signal with
default arguments).

Task-number: PYSIDE-462
Change-Id: I0fc0f37bc7539a1357b8d3cda3155403fca4a763
Reviewed-by: Christian Tismer <tismer@stackless.com>