Handle possible Python errors when invoking a global receiver callback
Previously the GlobalReceiverV2 instance did not check for the
occurrence of an error when calling its Python callback. This led to
to skipping all processing of QEvents as described in PYSIDE-478.

The fix is to check if a Python error ocurred, and print it (thus
handling it, and clearing the current active error).

Task-number: PYSIDE-360
Change-Id: Idc49dc6fc18da27e92043df8fbc46072efc73f72
Reviewed-by: Christian Tismer <tismer@stackless.com>