Remove memory leak in shiboken
There was a harder to find memory leak reported in early PySide 1 times
which was still valid in PySide2.

I used two scripts which were different by only one line and showed very
different memory behavior. With valgrind, I ran both scripts in parallel,
and after some tweaking, the valgrind logfiles of the two runs could be
compared.

The result was a clear indicator that some new call was never disposed of.
It turned out that the error was a mis-placed ‘delete’ in basewrapper.cpp
of shiboken2.

Task-number: PYSIDE-205
Change-Id: I1897ec4e75e3ec887716bdbe0f4487176530e03c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>