[Reland] Add QSharedPointer<QQuickItemGrabResult> to the type system
Also add test to check that calling QQuickItem::grabToImage() works
correctly.

The QSharedPointer type is added in the QtQuick typesystem file
and not in QtCore to minimize the surface of issues that might come up,
as well as because that's the only reasonable use case of QSharedPointer
in the public API so far.

The commit was missing in 5.9 branch after repo merge.

Task-number: PYSIDE-454
Change-Id: Ibb7775117ffb22ab91d2ac798ae13cc9e4856587
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>