Improve Qt initialization order hack
In CppGenerator::finishGeneration(), the code tried to reorder the
classlist such that the initialization of QMetaObject is written
before the initialization of QObject by re-inserting the QMetaObject
class entry in front of QObject. This does not consider the dependency
of QMetaObject::Connection on QMetaObject and worked only by
coincidence since the old parser did not recognize
QMetaObject::Connection as an inner class due to it being
forward-declared.

With the Clang-based parser, which recognizes this, the initialization
for QMetaObject::Connection would be called before the initialization
of QMetaObject, causing a crash.

Fix this by making it possible to pass additional dependencies
as pairs of QString to the topological sorting functions, which
then generate the correct sequence.

Task-number: PYSIDE-323
Change-Id: Ia915b47131d57e71df366876a1a9f317cfd8d497
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>