Define Q_QDOC while we have the old parser
Defining Q_QDOC enables some more modules, since it simplifies some
interfaces so the shiboken parser does not reject them.

This should be removed again, when we have a really good parser.
The feature was a bit forgotten. We needed to decide whether Q_QDOC is to
be set or not, and we decided “Yes, for Qt 5.6 we define it”.
This patch realizes that decision.

Change-Id: Ic2457e3e6791ded20d9e3a10778fef4f004ede43
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>