AbstractMetaBuilder::traverseFunctions() Split out duplicate removal
The removal of duplicate member function entries depended on the
fact that the more specific, out-of-line function definitions were
first in the list of functions returned by ScopeModelItem::functions().

This happened to work by coincidence since ScopeModelItem used a
QHash, which reversed the order returned by the parser.
When trying to replace the QHash by a QList in order to get
rid of the non-determinism,
TestTemplates::testContainerTypeIncompleteArgument() would fail
since it finds 2 entries for "method()".

Change-Id: I7c5a49bef8644d1f05e6a14295b304b1a9b8c46f
Reviewed-by: Christian Tismer <tismer@stackless.com>