Fix qstandarditemmodel_test.py not finding shiboken on macOS
When running make test inside the build folder of PySide2,
the mentioned test failed because it couldn't find shiboken2.
Make sure to import the module as standalone, so the test
succeeds.

Also make sure to specify the proper python paths, by computing
the paths to shiboken and pysidetest from the pyside build root
path, rather than from pysidetest build path, because on some
CMake implementations it evaluates to an empty path (probably
due to the fact that ${pysidetest_BINARY_DIR} is evaluated
before the pysidetest project is actually added).

Change-Id: Ibe86b0d7c2c66db708bba2ffc32b02e5c3e91c79
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>