Fix Inheritance of PySide
This patch is the third stab at fixing the inheritance of PySide.
The problem is now solved.

Inheritance was already correct when the functions were disjoint.
But as soon as a function existed in both a class and a sub-class,
only the class has won, and the sub-class was ignored.

This patch fixes that problem. There are 49 new functions, and
of course there is QApplication.palette() without an argument.

Function ‘searchForEnumScope’ was now corrected, so that the
work-around no longer was needed. I had to extend its search.

It turned out that the same fix for inheritance was also needed for
a correct listing of the allowed function signatures. I finally factored
out a new function that is used twice.

Exhaustive tests can be found in PySide with the same name.

Task-number: PYSIDE-331
Change-Id: If3d7d94f6ed351b5de03e26c9e3e65cf4a5f8ed7
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>