Fix Python extension suffix selection on Linux
This allows building Shiboken in any configuration (debug or release)
against any version of a Python interpreter (debug or release).

Previously specifying --debug forced the user to use a debug version
of the Python interpreter.

The patch also removes an unnecessary cmake file, which was used
in the past to find the location of Python libraries. This has
been handled by CMake itself for a while now.

Change-Id: I70edcdfee5fde246c1c41cd5b36353cc18aa9bdd
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Christian Tismer <tismer@stackless.com>