Fix Inheritance of PySide
These are the tests for the newly inherited functions.

By writing the tests, four new bugs have been found!
This told me that it is necessary to write a simple script that
systematically probes every generated function for basic functionality.

PySide2.QtPrintSupport.QPrintPreviewDialog().open() prints a newline.
Why is that?

Task-number: PYSIDE-331
Change-Id: Ib6780884dfb10cf43a36006c6cabef60b1e1750f
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>