Fix popenasync.py for non-Ascii characters
Some windows guy had some files with funny characters in the file names.
This caused a crash because no error handling was defined.
The script then got very confused and ran even into code which was for linux,
barfing about a missing fcntl module, and that created a total mess, because
the guy installed some funny fcntl module and got even worse errors.

The simple fix was to ignore the error, which made the guy very happy.
Happened with python3 on windows. Python2 does not have this issue.

I’m not sure if we should continue to use this module at all?

Task-number: PYSIDE-497
Change-Id: I48099a135702225c12cf7d8e62f058c50f612e5f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>