Fix Inheritance of PySide, Windows support
This patch is augmenting the general fix. We had not tested the case
with “#define AVOID_PROTECTED_HACK”. This setting can be used
under Linux and macOS, but is necessary under Windows, because the
recent MS compilers do not allow the setting “#define protected public”.

Therefore, certain special cases had to be circumvented by wrapper type casts,
and a few more header files were needed.

This patch was not tested on Windows, but built on macOS with
the mentioned setting.

Exhaustive tests can be found in PySide. They are now working as before.

Task-number: PYSIDE-500
Change-Id: I9c46d0f181f810879dc75326f2cf39fd2ec853f3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>