Rewrite CMake scripts and generate includes for global.h.in
The CMake files are repeating code three times, which should be avoided.
Also, the global.h.in contained too many unconditional includes, which becomes
relevant for the PySide 5.9 branch when missing includes are no longer ignored.
Instead of maintaining an always growing list of conditionals (did that first),
the needed includes are now computed by CMake, and the collected
includes are appended to pyside2_global.h .

Task-number: PYSIDE-507
Change-Id: I86f27d42c2d60d75ab4597e262e874c7186389c8
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>