Collect Qt packages earlier to support QML exceptions
When a80a6eb94433529a6d984e8ff22874f7dff48eea was introduced,
collection of Qt packages was moved from the top level CMake project,
to the PySide2 CMake subproject. This caused the issue that libpyside
could not find the QML private API classes (due to the packages not
being found in the project scope), leading to the failing test
javascript_exceptions.py.

This change moves the collection of the Qt packages back into the
top-level project, so that the found Qt packages are propagated to
all subprojects, including libpyside.

Change-Id: Ieb41a51cc40b11c9f9b81143bf1dbe7d9c97da0c
Reviewed-by: Christian Tismer <tismer@stackless.com>