Fix double inclusion of qsharedpointer_impl.h file
This is a workaround to allow building QSharedPointer bindings
with the clang parser.

In the previous parser the Q_QDOC define was set, and thus the
implementation header was never included. When clang is queried
to return the file which implements QSharedPointer, it will return
the innermost header file. Because the header doesn't have header
guards, the build will fail with redefinition errors.

Task-number: PYSIDE-454
Change-Id: I5628fe2bafe528e11076f344b18b463db90c5651
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>