Change parsing to #include the source header
shiboken is normally run on a global header that includes
all classes to be wrapped. However, when passing a header
that contains the class declaration, it happens that
the temporary copy of the header (being the location of the class
declaration) is stored as required #include
by the generators. Fix by using #include instead, so that the original
header from the command line is seen as location.

Task-number: PYSIDE-526
Change-Id: Ie30a8a2b21606c9f6ed5dda3447885576fdc4c2c
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>