Preserve symlinks instead of duplicating files
This change overrides some of the commands provided by distutils /
setuptools to preserve symlinks when copying directories. This is mostly
to decrease the time spent and amount of space used by copying into
the pyside_package, build/xxx and site-packages directories (copying
around WebEngine libraries takes a bit of time and space).

It is not useful for bdist_wheels or eggs, because Python's zip module
does not preserve symlinks when archiving the files.

Ultimately this will probably go away, once we figure out which if any
of the symlinks are needed.

Task-number: PYSIDE-495
Change-Id: I8766266c9ab51d610c9a5377618ab06020637d7d
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>