Allow building subset of Qt modules
Collect the list of essential and optional modules in list variables
and concatenate them to a list variable MODULES unless its value was
passed in on the command line with -D.

Remove the additional parameter indicating the module type to
COLLECT_MODULE_IF_FOUND() and let it determine that by checking
the essentials list instead.

Add a command line option --module-subset to setup.py which can be used
like "--module-subset=QtCore,QtGui,QtWidgets,QtTest" to specify the modules
to be built.

Change-Id: Ibb0fa16a8d9b0d7aeeaf8a8cfcbd50cb910ecc97
Reviewed-by: Christian Tismer <tismer@stackless.com>