Investigate the sporadic segfaults on RHEL
Lately, the RHEL platform tends to segfault between 0-4 times
in a test run.
I suspect that is related to some finalization code that is triggered
at a special time, due to a stressy situation. But we don't know
until we ask the CI system by printing a stack trace.

This patch prints a stack dump after a segfault happened.
The dump is only activated when the environment setting

    QTEST_ENVIRONMENT=ci

is true. (The above needs to be compared with strcmp or strstr)

Note that the tests _are not isolated_. I recognized that by the warnings
module that leaked between tests. The processes are shared, unless
you have setup and teardown functions.

There might for instance be some hidden refcount bug that triggers
only early enough in a stress situation.

Change-Id: Ibbc592e7bdcbdd3cdbc06f203be3e5fc1daaf774
Reviewed-by: Christian Tismer <tismer@stackless.com>