Create a function existence registry
With the signature module, it is now a straight forward task
to generate a registry of all known function signatures.
We check that these signatures all exist.

One file contains all signatures for one platform and version.

The test is only activated when run in the CI system.
An initial call creates the expected file as output and raises
an error. The result can then be picked up from the error log
and added to the repository.

Done: linux2 5.6.4
Done: darwin 5.6.4
Done: win32 5.6.4
Done: darwin 5.9.3
Done: linux2 5.9.3
Done: linux 5.9.3

Task-number: PYSIDE-510
Change-Id: I32dfd5fcd56ca8d91d48a81959cc762cd5340c68
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>