Create a function existence registry
With the signature module, it is now a straight forward task
to generate a registry of all known function signatures.
We check that these signatures all exist.

One file contains all signatures for one platform and version.

The test is only activated when run in the CI system.
An initial call creates the expected file as output and raises
an error. The result can then be picked up from the error log
and added to the repository.

The linux2 and linux platforms are now unified.

There will be a new version of testrunner.py which is more versatile.
In future, this teach-in process will be made much easier because
we will be able to view the initial versions without raising an error.

Done: linux 5.6.4
Done: darwin 5.6.4
Done: win32 5.6.4
Done: darwin 5.9.3
Done: linux 5.9.3
Done: win32 5.9.3

Task-number: PYSIDE-510
Change-Id: I4e3953f9d8e781ae170b455dc7884da9bf632f47
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>