testrunner 1: Turn into a package
Testrunner is a medium-large module that has a number of
different purposes. Since we want to improve the testing logic,
it is better to clean the module up before adding even more
code.
For compatibitity, the testrunner module will stay in place,
but redirect everything into the new "testing" package.

Task-number: PYSIDE-578
Change-Id: I3d14f3c4c421cdd2bb7777691b56677aa4786dd8
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>