testrunner 5: Improve the algorithm
Testrunner checks if it is running in COIN.
If so, it runs each tested project 5 times and reports an error if
3 errors were found in a test function and not blacklisted.

The time is measured, so we know if we can run all the tests five
times or if we can run only the failed tests.

At the moment, the option COIN_RERUN_FAILED_ONLY is activated
by default. We can configure it by setting to false.

Since this change turns the registry existence test into a flaky test
(te tests generates it's missing file in the first run),
the registry existence test is no more blacklisted.

We simulate our own tee command now with decorated output.

Task-number: PYSIDE-578
Change-Id: I6390cd50398a97a168c85c6bb778984c6a5b90fc
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>