Fix QtCore test bug829 to run on Windows
The test used Python functionality to create a temporary
file which caused an access error on Windows, apparently due to
the file still being open. Besides, the temporary file was leaked.

Rewrite using QTemporaryFile and make sure the file is closed
and the QTemporaryFile is deleted before using it and delete it
in __del__().

Task-number: PYSIDE-431
Change-Id: I5d63e76ead7169d9f3f2267ee53aa91b601a3968
Reviewed-by: Christian Tismer <tismer@stackless.com>