PEP 384-1: Estimate the problem size
Pep384 is quite restrictive and not at all implemented.

This is the beginning of a series of check-ins that should end in a
complete implementation of the restricted API.

The use of python.h has generally been changed to sbkpython.h .
This way we have a central entry point that defines completely
the usage of Python in PySide.

Right now it collects all problems to solve in the file pep384resolve.h .
When after some time this file becomes totally resolved and all rules
are checked again, then the implementation of the PEP should be ready.

Result:
There are 18 header files involved. These are
    longobject.h
    object.h
    pydebug.h
    unicodeobject.h
    bytesobject.h
    floatobject.h
    memoryobject.h
    tupleobject.h
    listobject.h
    methodobject.h
    descrobject.h
    pystate.h
    pythonrun.h
    abstract.h
    funcobject.h
    classobject.h
    code.h
    datetime.h

Task-number: PYSIDE-560
Change-Id: Ic66ccf27222aacc68fe2f63f4643ad7a6559e351