Enhance testrunner to ignore old test logs
Testrunner had a virtual problem:
When running testrunner without building the project again, and the test
skips some runs that were not skipped before, then it will fetch an old
test log from the former run.

Cure: We check all logfiles first. If a file exists and skip is not set, then we
remove the file. Then normal testing starts.

Change-Id: Ib91baca97d3b0a5c0b68a8022fca52202b20e862
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>