PEP 384-3: Replace builtin variables by Python calls
This patch contains more complicated fixes.
The variable Py_VerboseFlag is in pydebug.h, and because debugging
is not supported by the limited API, the Py_VerboseFlag is not accessible.
I think the Python developers were a bit too rigorous.
Anyway, this is quite cheap to fix and can be done without changes to
the source code:

Py_VerboseFlag gets replaced in a macro by a function call to
Py_GetVerboseFlag(). This is very fast due to caching.

The module datetime is replaces by Python in a similar way. It is not
optimized for speed. This is anyway mainly used for tests, only.

Solved:
    datetime.h
    pydebug.h
    floatobject.h
    tupleobject.h
    listobject.h

Unsolved:
    longobject.h
    object.h
    unicodeobject.h
    bytesobject.h
    memoryobject.h
    methodobject.h
    descrobject.h
    pystate.h
    pythonrun.h
    abstract.h
    funcobject.h
    classobject.h
    code.h

Task-number: PYSIDE-560
Change-Id: I5b074da32030aaf2a187d941df16afa2979b4b53