Improve logging of all relevant build folders that are written to
This change improves the output of which directories are used when a
build is done, as well as adds output for the correct final
installation destination (deduced automatically by setuptools, or
specified by --prefix option).

The change aims to reduce the usual confusion of figuring out which
files are written where.

Change-Id: I038be7bc657a7fa68e242c38076ffc4ba0548a0b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>