Improve libICU deployment on Linux
Previously the --standalone build process would download and extract
an archive of ICU libraries, regardless of which ICU Qt was built
against.

The build process will now detect which ICU libraries QtCore depends
on and copy the libraries over to the destintation libdir.

Something similar might be needed in the future for macOS and Windows.

Change-Id: I0db0c8c628d3c095a8a4a1e361f8fafe18da2ec3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>