PEP 384-7: Adjust signature.cpp to the limited API
This series of checkins will be ready for review after step 8.

Missing:
- object.h needs the heap type patch,
- We must implement pep384 as an option for pyside-setup,
- Windows is more rigorously removing types completely.
+ PYSIDE-593 has been fixed.

The signature module had some spots were it was quite incompatible with
PEP384. There could be almost all removed by using the fact that this is
Python 3 and we therefore have the __qualname__ attribute.

In the end, the only remaining wart was the usage of tp_dict, which I
replaced by a call to getattr("__dict__"), which gives us only a dictproxy.
I cheated and used the unpublished type 'mappingproxyobject'.

Instead of forcing a manual test of the validity of this breakage of
PEP384, I added code that computes the position of tp_dict and
tp_methods and ensures that they are exactly 4 pointers apart.

I hope that this is sufficient as a validation.

Furthermore unicodeobject.h was solved which was a bit tricky.

Solved:
    unicodeobject.h
    longobject.h
    descrobject.h
    pystate.h
    funcobject.h
    classobject.h
    abstract.h
    pythonrun.h
    code.h
    methodobject.h
    memoryobject.h
    bytesobject.h
    datetime.h
    pydebug.h
    floatobject.h
    tupleobject.h
    listobject.h

Unsolved:
    object.h

Task-number: PYSIDE-560
Task-number: PYSIDE-510
Change-Id: I670b423819595f02aa3c0af67da9fca56a6dd65c